Start consolidating utility functions.
Eventually they should end up in a separate file, and
probably a separate repo too. The need for _GNU_SOURCE
is unfortunate, not sure what can be done there. Maybe
convert to write()?
